refactor: move pkg/git to internal/git
